refactor: Split data tab link config file
